Change the title of the agreement
As mentioned in https://trello.com/c/NfJnwB1I/83-oct-10-session-microcopy-changes, we don't want this to sound like a contract. Also, add '(opens in new tab)' to links that open in a new tab. Seems only fair to warn users.
Add a line break after intro instructions so that the generated statement doesn't appear to be a continuation of the instructions.