Microcopy and content changes from Oct 10 research (#104)
* Add hint text to partner department question

From: https://trello.com/c/NfJnwB1I/83-oct-10-session-microcopy-changes
Provide example of a department name in hint text so the user realizes we're not asking for an initialism.

* Content changes from Oct 10 research session

Make changes based on analysis of October 10 research session as outlined in https://trello.com/c/NfJnwB1I/83-oct-10-session-microcopy-changes
Make the order of fields better match that of the actual form so that it's easier for future edits to use.
Add placeholders for a future template/agreement selection screen (`template-select.title` etc)

* Put the comma back in

* Change the title of the agreement

As mentioned in https://trello.com/c/NfJnwB1I/83-oct-10-session-microcopy-changes, we don't want this to sound like a contract. Also, add '(opens in new tab)' to links that open in a new tab. Seems only fair to warn users.
Add a line break after intro instructions so that the generated statement doesn't appear to be a continuation of the instructions.

* Remove different comma

* Add intro above questions

Add intro text between title and h2, explaining briefly what the form does with input text.