fix: use node:12.22.12 base image
Update the Dockerfile used by Heroku to use the node:12.22.12 base image
so we can pin to the expected Node.js 12.x required by the project.  As
part of this change, we now need to manually install pandoc.

We moved away from using pandoc/core as the base image because alpine
is no longer providing Node.js 12.x packages.