fix: docx file generation and download
Update the Lambda Terraform to give pandoc time to generate
the docx file and base64 encode the response.

Update the app to serve the agreement summary response
from within the nodePandoc callback.  This forces the Lambda
to generate the docx file before terminating.  Also adds a
`/access/:fileName` route to stream the generated file from
the `/mnt/access` folder.